ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.24.4

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Optional Dependencies  
# Note: vLLM not available on Windows. Using transformers with FP16 instead.
# openai>=1.0.0  # For fallback
# redis>=5.0.0  # For caching
# psycopg2-binary>=2.9.9  # For PostgreSQL

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

